# apps/web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
      "AddFriend": {
        "additionalProperties": false,
        "properties": {
          "friend_code": {
            "title": "Friend Code",
            "type": "string"
          }
        },
        "required": [
          "friend_code"
        ],
        "title": "AddFriend",
        "type": "object"
      },
...
      "ExperimentRequest": {
        "additionalProperties": false,
        "properties": {
          "cue": {
            "default": "banana",
            "enum": [
              "banana",
              "apple_vinegar",
              "yeast",
              "grape",
              "blue_light"
            ],
            "title": "Cue",
            "type": "string"
          },
          "kind": {
            "const": "odor_choice",
            "default": "odor_choice",
            "title": "Kind",
            "type": "string"
          },
          "trials": {
            "default": 100,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Trials",
            "type": "integer"
          }
        },
        "title": "ExperimentRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SendGift": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Amount",
            "type": "integer"
          },
          "material": {
            "title": "Material",
            "type": "string"
          }
        },
        "required": [
          "material",
          "amount"
        ],
        "title": "SendGift",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Tsuyu Labo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/flies/{fly_id}/behavior": {
      "get": {
        "operationId": "behavior_v1_flies__fly_id__behavior_get",
        "parameters": [
          {
            "in": "path",
            "name": "fly_id",
            "required": true,
            "schema": {
              "title": "Fly Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scenario",
            "required": false,
            "schema": {
              "default": "rest",
              "pattern": "^(rest|sugar|bitter|sugar\\+bitter|looming|light_left|light_right|antenna_touch|liked_odor|disliked_odor)$",
              "title": "Scenario",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "number"
                  },
                  "title": "Response Behavior V1 Flies  Fly Id  Behavior Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Behavior"
      }
    },
    "/v1/flies/{fly_id}/experiments": {
      "post": {
        "operationId": "experiment_v1_flies__fly_id__experiments_post",
        "parameters": [
          {
            "in": "path",
            "name": "fly_id",
            "required": true,
            "schema": {
              "title": "Fly Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Experiment V1 Flies  Fly Id  Experiments Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Experiment"
      }
    },
    "/v1/friends": {
      "get": {
        "operationId": "list_friends_v1_friends_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Friends V1 Friends Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Friends"
      },
      "post": {
        "operationId": "add_friend_v1_friends_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFriend"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Friend V1 Friends Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Friend"
      }
    },
    "/v1/friends/{friend_id}": {
      "delete": {
        "operationId": "remove_friend_v1_friends__friend_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "friend_id",
            "required": true,
            "schema": {
              "title": "Friend Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Remove Friend V1 Friends  Friend Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Friend"
      }
    },
    "/v1/friends/{friend_id}/gift": {
      "post": {
        "operationId": "gift_v1_friends__friend_id__gift_post",
        "parameters": [
          {
            "in": "path",
            "name": "friend_id",
            "required": true,
            "schema": {
              "title": "Friend Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendGift"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Gift V1 Friends  Friend Id  Gift Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Gift"
      }
    },
    "/v1/friends/{friend_id}/lab": {
      "get": {
        "operationId": "lab_v1_friends__friend_id__lab_get",
        "parameters": [
          {
            "in": "path",
            "name": "friend_id",
            "required": true,
            "schema": {
              "title": "Friend Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Lab V1 Friends  Friend Id  Lab Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lab"
      }
    },
    "/v1/friends/{friend_id}/like": {
      "post": {
        "operationId": "like_v1_friends__friend_id__like_post",
        "parameters": [
          {
            "in": "path",
            "name": "friend_id",
            "required": true,
            "schema": {
              "title": "Friend Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Like V1 Friends  Friend Id  Like Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Like"
      }
    },
...
    "/v1/sleep/end": {
      "post": {
        "operationId": "end_v1_sleep_end_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response End V1 Sleep End Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End"
      }
    },
...
  }
}